clean up debug statements to a minimum

+++ keystone-sbi.h
@@ -37,7 +37,7 @@ struct keystone_sbi_create_t
   struct runtime_params_t params;
   
   // policy registration
-  int cycles_per_epoch;
+  uint64_t cycles_per_epoch;
 };
 
 struct sbiret sbi_sm_create_enclave(struct keystone_sbi_create_t* args);
